cleaned code for Aster
